refactor!: Refactor into component
