+++ messenger/component/sender_action/test_sender_action.py
@@ -2,7 +2,7 @@ import os
 
 import pytest
 
-from messenger.sender_action import service
+from messenger.component import sender_action
 
 
 @pytest.fixture()
@@ -13,9 +13,9 @@ def page_scoped_id():
 @pytest.mark.parametrize(
     "action",
     [
-        service.mark_seen,
-        service.typing_on,
-        service.typing_off,
+        sender_action.mark_seen,
+        sender_action.typing_on,
+        sender_action.typing_off,
     ],
     ids=["mark_seen", "typing_on", "typing_off"],
 )
