+++ messenger/controller.py
@@ -1,8 +1,6 @@
 from flask import Blueprint, request
 
-from messenger.verify.service import verify_service
-from messenger.messages.router import message_response_router
-from messenger.messaging_referrals import service as messaging_referrals_service
+from messenger import verify, messages, messaging_referrals
 
 messenger_controller = Blueprint("messenger", __name__)
 
@@ -11,7 +9,7 @@ messenger_controller = Blueprint("messenger", __name__)
 def verify_controller():
     query = request.args
 
-    response = verify_service(query)
+    response = verify.verify_service(query)
     return response if response else ("ok", 200)
 
 
@@ -30,9 +28,9 @@ def handle_entry(entry):
     messaging = entry["messaging"].pop()
 
     if messaging.get("message"):
-        handler = message_response_router
+        handler = messages.handler
     elif messaging.get("messaging_referrals"):
-        handler = messaging_referrals_service.handler
+        handler = messaging_referrals.handler
     else:
         handler = lambda _: True
