--- messenger/messages/debug/service.py
@@ -1,8 +1,0 @@
-import json
-
-from messenger.messages.repository import send_message_response
-
-
-def debug_send_message_response(messaging):
-    send_message_response(messaging["sender"]["id"], {"text": json.dumps(messaging)})
-    return True
